+++ Changes.md
@@ -49,6 +49,7 @@ Improvements
   - Matched DelightOptions default values for `oversampling` and `shadingSamples` to 3Delight's own default values.
 - GraphEditor : Improved logic used to connect a newly created node to the selected nodes.
 - ScenePlug, ImagePlug : Child plugs are now serialisable. Among other things, this enables them to be driven by expressions (#3986).
+- Premultiply : Added `useDeepVisibility` plug, which weights samples according to their visibility based on the opacity of samples in front.
 
 Fixes
 -----